Add "identify owner" to the options
This patch adds the selection `I` for "(I)dentify owner" to the list
of options that can be executed on a found rpmnew/rpmsave file. This
calls `rpm -qf` to identify the source package the file belongs to.

This option can be usefull to check if the file was created by an
update that you are currently working on or if the file is a
leftover of a previous update. If multiple persons are adminitrating
a system this can be a valuable information.