keep permissions of the original File
Changed the `mv`-command to `cp`-command to preserve the
File-Permissions of the original File. Using the `mv`-command
kept the permissions from the rpmsave/rpmnew-file wich
maybe installed with mode `600` and/or `root:root`.
This will mess up the Permissions of the desitination File.

This Patch uses `cp`-command wich keeps the permissions of the
original File.